feat(KL-087): KarmoMap 세계관 관계도 위젯 scaffold + lib/graph 추출
단위 0 — lib 추출:
- src/lib/graph/spec.ts (타입 정의, Tauri 의존 제거)
- src/lib/graph/adapter.ts (GraphPersistAdapter 인터페이스 + NullAdapter)
- src/lib/graph/canvas.ts (GraphCanvas 이주, persistAdapter/kindColors 주입)
- cockpit/graph-tauri-adapter.ts (CockpitTauriAdapter 신설)
- cockpit/graph-canvas.ts → lib re-export shim
- cockpit/cockpit.ts → CockpitTauriAdapter 사용

단위 1 — KarmoMap 위젯:
- widgets/karmomap/kinds.ts (Character/Place/Item/Event/Concept + 엣지 5종)
- widgets/karmomap/local-storage-adapter.ts (karmomap.* localStorage)
- widgets/karmomap/karmomap.ts (렌더러 + 노드추가/삭제/엣지연결/JSON export-import)
- widgets/karmomap/index.ts (IIFE + Toolbox.register)
- widgets-lazy-meta.ts (karmomap 메타 등록)
- build.mjs (karmomap 엔트리 추가)

behavior_verified = 사용자 로컬 동작 확인 후 (autopilot Draft PR)
